backend/requirements.txt
opencv-python
mediapipe
python-socketio
aiohttp
uvicorn
